chore: create helpers
